backend/records/serializers/custom_token_serializer.py
from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
from rest_framework import serializers
from django.contrib.auth import authenticate
from django.utils.translation import gettext_lazy as _

class CustomTokenObtainPairSerializer(TokenObtainPairSerializer):
    username_field = 'email'

    def validate(self, attrs):
        email = attrs.get("email")
        password = attrs.get("password")

        if email and password:
            user = authenticate(request=self.context.get('request'), username=email, password=password)

            if not user:
                raise serializers.ValidationError(_("メールアドレスまたはパスワードが正しくありません"))
        else:
            raise serializers.ValidationError(_("メールアドレスとパスワードを入力してください"))

        data = super().validate(attrs)
        return data